refactor(): move user dto & responses to it's package in schemas
