+++ app/api/change_password.py
@@ -8,8 +8,8 @@ from sqlalchemy.orm import Session
 from app.db.database import get_session
 from app.enums.app_enums import HandlerTypes
 from app.exceptions.user_exceptions import UserWrongPasswordException, UserNotFoundException
-from app.schemas.responses.user_account_response import UserAccountResponse
-from app.schemas.user_change_password import UserChangePassword
+from app.schemas.user.user_account_response import UserAccountResponse
+from app.schemas.user.user_change_password import UserChangePassword
 from app.services.user_service import UserService
 
 router = APIRouter()
