+++ app/api/delete_user.py
@@ -1,3 +1,4 @@
+from http import HTTPStatus
 from uuid import UUID
 
 from fastapi import APIRouter, HTTPException
@@ -5,6 +6,7 @@ from fastapi.params import Depends
 from sqlalchemy.orm import Session
 
 from app.db.database import get_session
+from app.enums.app_enums import HandlerTypes
 from app.exceptions.user_exceptions import UserNotFoundException
 from app.services.user_service import UserService
 
@@ -12,7 +14,14 @@ router = APIRouter()
 
 
 @router.delete(
-    path="/user/{user_id}"
+    path="/user/{user_id}",
+    tags=[HandlerTypes.USER],
+    status_code=HTTPStatus.NO_CONTENT,
+    summary="Удаление пользователя",
+    description="Удаление аккаунта. Пользователь должен быть авторизован. Удалить можно только свой аккаунт",
+    responses={
+        HTTPStatus.NO_CONTENT: {"description": "Пользователь успешно удален"}
+    }
 )
 async def delete_user(
         user_id: UUID,
