+++ app/api/upload_document.py
@@ -7,8 +7,8 @@ from app.db.database import get_session
 from app.dependencies.auth import get_current_user
 from app.enums.app_enums import HandlerTypes
 from app.exceptions.collection_exceptions import BaseCollectionNotFoundException
-from app.schemas.document_upload import DocumentUploadResponse
-from app.schemas.user_dto import UserDTO
+from app.schemas.document.document_upload_response import DocumentUploadResponse
+from app.schemas.user.user_dto import UserDTO
 from app.services.collection_service import CollectionService
 from app.services.document_service import DocumentService
