--- app/schemas/user_file_dto.py
@@ -1,11 +1,0 @@
-from datetime import datetime
-
-from pydantic import BaseModel
-
-
-class UserFileDTO(BaseModel):
-    file_name: str
-    file_size: int
-    load_datetime: datetime
-    user: str
-    words: list[str]
